votebot/requirements.txt
PyYAML==3.11
requests==2.9.1
six==1.10.0
slackclient==1.0.0
tinydb==3.1.3
ushlex==0.99
websocket-client==0.35.0
wheel==0.24.0
